Update _helper_functions.py
added int() to make sure factorial function is executable

+++ _helper_functions.py
@@ -66,7 +66,7 @@ def R(n, m, rho):
     else:
         for k in range(int((n - m) / 2) + 1):
             coef = ((-1) ** k * factorial(n - k) /
-                    (factorial(k) * factorial((n + m) / 2. - k) * factorial((n - m) / 2. - k)))
+                    (factorial(int(k)) * factorial(int((n + m) / 2. - k)) * factorial(int((n - m) / 2. - k))))
             output += coef * rho ** (n - 2 * k)
         return output
